pyproject.toml
[project]
name = "inflation_tracker"
version = "0.5.0"
description = "Трекер для отслеживания инляции"
readme = "README.md"
classifiers = ["Programming Language :: Python :: 3.9"]
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.0,<1.17",
    "sqlalchemy>=2.0.45",
    "prettytable>=3.10.0",
    "python-dotenv>=1.0.0",
    "pandas==2.3.3",
    "matplotlib==3.9.4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.11",
]
testing = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
gui = [
  "PyQt6>=6.6",
]
packaging = [
    "pyinstaller>=6.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[tool.ruff]
line-length = 79
target-version = "py39"
src = ["app"]
exclude = [
    ".git",
    ".ruff_cache",
    ".pytest_cache",
    "venv",
    ".venv",
    "alembic/versions",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 60
line-ending = "lf"
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C90", "UP", "SIM", "RET"]
ignore = [
    "E265",
    "F811",
    "D100",
    "D105",
    "D104",
    "D203",
    "D211",
    "D213",
    "N818",
    "W505",
    "ANN401",
    "A005",
    "A002",
    "RET504",
    "UP045",
    "N806",
    "UP006",
    "UP035",
    "N805",
    "B904",
    "N802",
    "B008",
    "C901"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "E401"]
"tests/**" = ["D", "ANN"]
"app/cli/**" = ["D"]
